+++ cli/src/commands/mod.rs
@@ -1,8 +1,10 @@
 pub mod avatar;
 pub mod dev;
+pub mod engine;
 pub mod game;
 pub mod image;
 pub mod init;
 pub mod ns;
 pub mod site;
+pub mod token;
 pub mod version;