+++ scripts/copy_libs.sh
@@ -1,0 +1,10 @@
+# Copy libraries
+rm -rf lib/rivet-api/
+cp -r ../rivet/gen/openapi/internal/rust/ lib/rivet-api/
+
+# HACK: Modify libraries to disallow unknown fields in config
+find lib/rivet-api -name "*.rs" -exec sed -i 's/\(#\[derive.*Deserialize.*\]\)/\1\n#[serde(deny_unknown_fields)]/g' {} \;
+
+(cd lib/rivet-api && cargo fmt)
+
+