+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Unreal helper in `rivet init`
+- Installer for the Unreal Engine plugin with `rivet unreal install-plugin` or `rivet init --unreal`
+
+### Changed
+
+- Renamed `rivet.version.toml` to `rivet.toml`. All changes are backwards compatible.
+
 ## [v0.0.51] - 2023-04-26
 
 ### Fixed
@@ -37,4 +48,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update dependency: `rivet-api`
 - Upgrade dependency: `tokio 1.27`
 - Removed unnecessary feature flags from `tokio`
-